+++ src/components/courses/baseCourse.vue
